move input to its own file, add constructor for icontrolpad
