+++ cmd/nes/main.go
@@ -48,84 +48,6 @@ func setupAudio(sampleRate float32) (sdl.AudioDeviceID, error) {
     return device, err
 }
 
-type SDLButtons struct {
-}
-
-func (buttons *SDLButtons) Get() nes.ButtonMapping {
-    mapping := make(nes.ButtonMapping)
-
-    keyboard := sdl.GetKeyboardState()
-    mapping[nes.ButtonIndexA] = keyboard[sdl.SCANCODE_A] == 1
-    mapping[nes.ButtonIndexB] = keyboard[sdl.SCANCODE_S] == 1
-    mapping[nes.ButtonIndexSelect] = keyboard[sdl.SCANCODE_Q] == 1
-    mapping[nes.ButtonIndexStart] = keyboard[sdl.SCANCODE_RETURN] == 1
-    mapping[nes.ButtonIndexUp] = keyboard[sdl.SCANCODE_UP] == 1
-    mapping[nes.ButtonIndexDown] = keyboard[sdl.SCANCODE_DOWN] == 1
-    mapping[nes.ButtonIndexLeft] = keyboard[sdl.SCANCODE_LEFT] == 1
-    mapping[nes.ButtonIndexRight] = keyboard[sdl.SCANCODE_RIGHT] == 1
-
-    return mapping
-}
-
-type SDLJoystickButtons struct {
-    joystick *sdl.Joystick
-}
-
-func OpenJoystick(index int) (SDLJoystickButtons, error){
-    joystick := sdl.JoystickOpen(index)
-    if joystick == nil {
-        return SDLJoystickButtons{}, fmt.Errorf("Could not open joystick %v", index)
-    }
-
-    return SDLJoystickButtons{joystick: joystick}, nil
-}
-
-func (joystick *SDLJoystickButtons) Close(){
-    joystick.joystick.Close()
-}
-
-func (joystick *SDLJoystickButtons) Get() nes.ButtonMapping {
-    mapping := make(nes.ButtonMapping)
-
-    mapping[nes.ButtonIndexA] = joystick.joystick.Button(12) == 1
-    mapping[nes.ButtonIndexB] = joystick.joystick.Button(13) == 1
-    mapping[nes.ButtonIndexSelect] = joystick.joystick.Button(8) == 1
-    mapping[nes.ButtonIndexStart] = joystick.joystick.Button(9) == 1
-    mapping[nes.ButtonIndexUp] =  joystick.joystick.Button(0) == 1
-    mapping[nes.ButtonIndexDown] = joystick.joystick.Button(3) == 1
-    mapping[nes.ButtonIndexLeft] = joystick.joystick.Button(2) == 1
-    mapping[nes.ButtonIndexRight] =  joystick.joystick.Button(1) == 1
-
-    return mapping
-}
-
-type CombineButtons struct {
-    Buttons []nes.HostInput
-}
-
-func MakeCombineButtons(input1 nes.HostInput, input2 nes.HostInput) CombineButtons {
-    return CombineButtons{
-        Buttons: []nes.HostInput{input1, input2},
-    }
-}
-
-func combineMapping(input1 nes.ButtonMapping, input2 nes.ButtonMapping) nes.ButtonMapping {
-    out := make(nes.ButtonMapping)
-    for _, button := range nes.AllButtons() {
-        out[button] = input1[button] || input2[button]
-    }
-
-    return out
-}
-
-func (combine *CombineButtons) Get() nes.ButtonMapping {
-    var mapping nes.ButtonMapping
-    for _, input := range combine.Buttons {
-        mapping = combineMapping(mapping, input.Get())
-    }
-
-    return mapping
-}
 
 func stripExtension(path string) string {
     extension := filepath.Ext(path)
@@ -379,7 +301,8 @@ func RunNES(path string, debug bool, maxCycles uint64, windowSizeMultiple int, r
 
     var joystickInput nes.HostInput
     if sdl.NumJoysticks() > 0 {
-        input, err := OpenJoystick(0)
+        // input, err := OpenJoystick(0)
+        input, err := MakeIControlPadInput(0)
         if err == nil {
             defer input.Close()
             joystickInput = &input
